Officially fixed avatars, fixed possible friend request remote deletion

+++ VocabCrack-iOS/FriendsController.swift
@@ -148,7 +148,12 @@ class FriendsController: TableDataReceiver
                 }
                 
                 updateModeTitle()
-                Handlers.friendHandler.deleteFriend(username, type: type)
+                
+                if remoteDelete
+                {
+                    Handlers.friendHandler.deleteFriend(username, type: type)
+                }
+                
                 tableView.deleteRowsAtIndexPaths([indexPath], withRowAnimation: .Fade)
             }
         }
